requirements.txt
librosa==0.11.0
pandas==2.2.3
numpy==2.2.6
torch==2.7.0
torchaudio==2.7.0
yt-dlp==2025.5.22
python-dotenv==1.1.0
spotipy==2.25.1
flask==3.1.1